IN-456: Add PHP linter
Also, in order to make it easier to add the PHP linter, the way the workflow works changed a little bit.

Now, instead running it inside the hadolint container, we run things on top of ubuntu which already comes installed with most of the things we need.

The Coder standard being followed is PSR-12. Since this is not a CiviCRM, Drupal or Symfony project, it doesn't make sense to follow any of their code standards, so PSR-12 was chosen for being the most generic and recent one available.