IN-476: Add support for PHP 5.6
Some of our projects haven't been updated to be compatible with PHP 7.2 yet,
so the action also needs to work with PHP 5.6.

The main piece of work to make that happen was actually done in the
`compucorp/civicrm-buildkit` image, which now offers php 5.6 variants.

With that done, we've introduced an `ARG` in the `Dockerfile` which allows
to specify which variation of the image we want to use. The default is
`7.2`, meaning that `compucorp/civicrm-buildkit` with php 7.2 will be used.

This change, however, forced us to also change how the image is build. Up
until now, we were using as `uses: compucorp/drupal7-security-updates`.
This would make Github Actions to clone the repo and build the image itself.
Unfortunately, it's not possible to pass build params to this process, which
means it would always build an PHP 7.2 image. To avoid this limitation, we
will now start pushing images to docker hub. This way, to tell Github to
download them from the registry instead of trying to build them itself.

This requires a few changes in how the action works though. Previously,
we would pass the `GITHUB_TOKEN` as an input to the action. However, actions
pulled from docker hub don't support that, so now we need to pass the token
via an Environment variable.

The `build-pull-request.php` script had to ve updated to be
compatible with PHP 5.6 as well.

A sample `civicrm.settings.php` was added to ensure the action can be used
on projects that also depend on civi, as during the tests it was found that
drush will warning about this file being missed in civicrm project (it would
not fail though).