IN-485: Ignore .gitignore changes
When updating Drupal core, the `.gitignore` file in the root folder might end-up being
updated as well. That's because Drupal core packages are shipped with a default
`.gitignore` file. As a result, this will overwrite any project specific
customizations that we might have end-up adding to the file.

To avoid that, the script will now check if the security updated changed the
`.gitignore` file and will revert the changes in case there are any.

The script will also now remove the `settings.php` and `civicrm.settings.php` file it
adds. This was done to avoid the files being committed to the security updates branch
in case the `.gitignore` file inside the repo is not set to ignore them.